Add ActionsMenu component
